% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ho Ho Ho.R
\name{ho_ho_ho}
\alias{ho_ho_ho}
\title{Plays jingle bells AND plots a Christmas tree in the plot pane!}
\usage{
ho_ho_ho()
}
\description{
Plays jingle bells AND plots a Christmas tree in the plot pane!
}
\author{
Ben Cuff
}
