% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jingle Bells.R
\name{jingle_bells}
\alias{jingle_bells}
\title{Plays jingle bells!}
\usage{
jingle_bells(vol = 0.1)
}
\arguments{
\item{vol}{The audio volume, from 0 (silent) to 1 (loud)}
}
\description{
Plays jingle bells!
}
\author{
Ben Cuff
}
